cast output as int to fix build error

+++ tensorflow/lite/micro/kernels/elementwise.cc
@@ -240,16 +240,18 @@ inline T AbsEvalQuantized(TfLiteContext* context, TfLiteNode* node, T i) {
 
   const int32_t value = std::abs(i - op_data->input_offset);
   if (!op_data->needs_rescale) {
-    return static_cast<T>(std::min(
-        std::max(value + op_data->output_offset, static_cast<int>(kMin)),
-        static_cast<int>(kMax)));
+    return static_cast<T>(
+        std::min(std::max(static_cast<long int>(value + op_data->output_offset),
+                          static_cast<long int>(kMin)),
+                 static_cast<long int>(kMax)));
   }
 
   const int32_t output = tflite::MultiplyByQuantizedMultiplier(
                              value, op_data->multiplier, op_data->shift) +
                          op_data->output_offset;
-  return static_cast<T>(std::min(std::max(output, static_cast<int>(kMin)),
-                                 static_cast<int>(kMax)));
+  return static_cast<T>(std::min(
+      std::max(static_cast<long int>(output), static_cast<long int>(kMin)),
+      static_cast<long int>(kMax)));
 }
 
 template <typename T>
@@ -274,8 +276,9 @@ inline T RsqrtEvalQuantized(TfLiteContext* context, TfLiteNode* node, T i) {
       tflite::MultiplyByQuantizedMultiplier(data, op_data->multiplier,
                                             op_data->shift - kShift) +
       op_data->output_offset;
-  return static_cast<T>(std::min(std::max(output, static_cast<int>(kMin)),
-                                 static_cast<int>(kMax)));
+  return static_cast<T>(std::min(
+      std::max(static_cast<long int>(output), static_cast<long int>(kMin)),
+      static_cast<long int>(kMax)));
 }
 
 template <typename T>
